renamed paramValues.h to valueContainers.h
